requirements.txt
discord.py
astral
requests
requests-cache
feedparser
asyncpraw
asyncprawcore
ollama
